+++ resolvers/deno/clippy.toml
@@ -1,0 +1,53 @@
+disallowed-methods = [
+  { path = "std::env::current_dir", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::Path::canonicalize", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::Path::is_dir", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::Path::is_file", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::Path::is_symlink", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::Path::metadata", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::Path::read_dir", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::Path::read_link", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::Path::symlink_metadata", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::Path::try_exists", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::PathBuf::exists", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::PathBuf::canonicalize", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::PathBuf::is_dir", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::PathBuf::is_file", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::PathBuf::is_symlink", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::PathBuf::metadata", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::PathBuf::read_dir", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::PathBuf::read_link", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::PathBuf::symlink_metadata", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::PathBuf::try_exists", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::env::set_current_dir", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::env::temp_dir", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::canonicalize", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::copy", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::create_dir_all", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::create_dir", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::DirBuilder::new", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::hard_link", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::metadata", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::OpenOptions::new", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::read_dir", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::read_link", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::read_to_string", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::read", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::remove_dir_all", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::remove_dir", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::remove_file", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::rename", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::set_permissions", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::symlink_metadata", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::fs::write", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::Path::canonicalize", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "std::path::Path::exists", reason = "File system operations should be done using DenoResolverFs trait" },
+  { path = "url::Url::to_file_path", reason = "Use deno_path_util instead so it works in Wasm" },
+  { path = "url::Url::from_file_path", reason = "Use deno_path_util instead so it works in Wasm" },
+  { path = "url::Url::from_directory_path", reason = "Use deno_path_util instead so it works in Wasm" },
+]
+disallowed-types = [
+
+  # todo(dsherret): consider for the future
+  # { path = "std::sync::Arc", reason = "use crate::sync::MaybeArc instead" },
+]
