+++ resolvers/deno/lib.rs
@@ -1,3 +1,5 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 
+pub mod fs;
+pub mod npm;
 pub mod sloppy_imports;
