+++ resolvers/deno/npm/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
+
+mod byonm;
+mod local;
+
+pub use byonm::ByonmNpmResolver;
+pub use byonm::ByonmNpmResolverCreateOptions;
+pub use local::normalize_pkg_name_for_node_modules_deno_folder;
