+++ resolvers/node/clippy.toml
@@ -42,6 +42,9 @@ disallowed-methods = [
   { path = "std::fs::write", reason = "File system operations should be done using NodeResolverFs trait" },
   { path = "std::path::Path::canonicalize", reason = "File system operations should be done using NodeResolverFs trait" },
   { path = "std::path::Path::exists", reason = "File system operations should be done using NodeResolverFs trait" },
+  { path = "url::Url::to_file_path", reason = "Use deno_path_util instead so it works in Wasm" },
+  { path = "url::Url::from_file_path", reason = "Use deno_path_util instead so it works in Wasm" },
+  { path = "url::Url::from_directory_path", reason = "Use deno_path_util instead so it works in Wasm" },
 ]
 disallowed-types = [
   { path = "std::sync::Arc", reason = "use crate::sync::MaybeArc instead" },
